# services/monolith/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Kvasir API
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value
...
paths:
...
  /{podId}/changes:
    get:
      tags:
      - Change History Api
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/ld+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChangeReport"
    post:
      tags:
      - Knowledge Graph API
      summary: Perform mutations on the KG.
      description: "Post a change request, containing the requested mutations, to\
        \ the inbox of the specified pod."
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/ld+json:
            schema:
              $ref: "#/components/schemas/ChangeRequestInput"
      responses:
        "202":
          description: Change request accepted.
  /{podId}/changes/{changeId}:
    get:
      tags:
      - Change History Api
      parameters:
      - name: changeId
        in: path
        required: true
        schema:
          type: string
      - name: podId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/ld+json:
              schema:
                $ref: "#/components/schemas/ChangeReport"
  /{podId}/changes/{changeId}/records:
    get:
      tags:
      - Change History Api
      parameters:
      - name: changeId
        in: path
        required: true
        schema:
          type: string
      - name: podId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/ld+json:
              schema:
                $ref: "#/components/schemas/ChangeRecords"
...
  /{podId}/slices/{sliceId}/changes:
    post:
      tags:
      - Knowledge Graph API
      summary: Perform mutations on a specific slice of the KG.
      description: "Post a change request, containing the requested mutations, to\
        \ a slice inbox of the specified pod."
      parameters:
      - name: podId
        in: path
        required: true
        schema:
          type: string
      - name: sliceId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/ld+json:
            schema:
              $ref: "#/components/schemas/ChangeRequestInput"
      responses:
        "202":
          description: Change request accepted.
...
components:
  schemas:
    Assertion:
      required:
      - '@type'
      - kss:query
      type: object
      properties:
        '@type':
          description: The type of the assertion.
          enum:
          - https://kvasir.discover.ilabt.imec.be/vocab#AssertEmptyResult
          - https://kvasir.discover.ilabt.imec.be/vocab#AssertNonEmptyResult
          type: string
          example: kss:AssertEmptyResult
        kss:query:
          description: The GraphQL query string to be executed.
          type: string
          example: "{ id ex_givenName(_: \"Bob\") }"
    ChangeRecords:
      required:
      - context
      - id
      - timestamp
      type: object
      properties:
        context:
          type: object
          additionalProperties: {}
        id:
          type: string
        timestamp:
          $ref: "#/components/schemas/Instant"
        deleted: {}
        inserted: {}
    ChangeReport:
      required:
      - id
      - podId
      - timestamp
      - resultCode
      type: object
      properties:
        id:
          type: string
        podId:
          type: string
        timestamp:
          $ref: "#/components/schemas/Instant"
        resultCode:
          $ref: "#/components/schemas/ChangeResultCode"
        sliceId:
          type: string
          nullable: true
        nrOfInserts:
          format: int64
          type: integer
        nrOfDeletes:
          format: int64
          type: integer
        errorMessage:
          type: string
          nullable: true
    ChangeRequestInput:
      required:
      - '@context'
      - kss:assert
      - kss:insert
      type: object
      properties:
        '@context':
          description: The JSON-LD context for the change request.
          type: object
          additionalProperties: {}
          example:
            kss: http://kvasir.discover.ilabt.imec.be/vocab#
            ex: http://example.org/
        kss:assert:
          description: List of assertions to be checked before applying the change
            request.
          type: array
          items:
            $ref: "#/components/schemas/Assertion"
        kss:with:
          description: Optional GraphQL query where matches are required to be found
            for the change request to be applied. Results are bound to the field names
            in the query and can be used in the insert and delete operations (via
            templates).
          type: string
          example: "{ id ex_givenName(_: \"Bob\") }"
          nullable: true
        kss:insert:
          description: "List of triples to be inserted, or a [JSONata](https://jsonata.org)\
            \ template string to be applied to the results of the with-clause."
          type: array
          items: {}
          example:
          - '@id': ex:123
            ex:givenName: Bob
    ChangeResultCode:
      enum:
      - COMMITTED
      - ASSERTION_FAILED
      - NO_MATCHES
      - TOO_MANY_MATCHES
      - VALIDATION_ERROR
      - INTERNAL_ERROR
      type: string
    Instant:
      format: date-time
      type: string
      example: 2022-03-10T16:15:50Z
...